MPG Driver is now partially working

+++ mpg/games/mpg.py
@@ -140,11 +140,11 @@ The second player is called the Min player, his goal is to minimize the mean-pay
 
     @property
     def weights_matrix(self):
-        return nx.to_numpy_array(self, weight="weight")
+        return nx.to_numpy_array(self, weight="weight",dtype=np.float32)
 
     @property
     def adjacency_matrix(self):
-        return nx.to_numpy_array(self, weight=None)
+        return nx.to_numpy_array(self, weight=None,dtype=np.float32)
 
     @property
     def tensor_representation(self):
